+++ src/extraction_lambda/extraction/extractor.py
