+++ src/secret_manager/password_manager.py
@@ -1,7 +1,7 @@
-from src.secret_manager.create_entry import create_entry
-from src.secret_manager.retrieve_entry import retrieve_entry
-from src.secret_manager.delete_entry import delete_entry
-from src.secret_manager.list_entries import list_secrets
+from src.extraction_lambda.secret_manager.create_entry import create_entry
+from src.extraction_lambda.secret_manager.retrieve_entry import retrieve_entry
+from src.extraction_lambda.secret_manager.delete_entry import delete_entry
+from src.extraction_lambda.secret_manager.list_entries import list_secrets
 
 prompt = "Please specify [e]ntry, [r]etrieval, [d]eletion, [l]isting or e[x]it: \n"
 
@@ -31,7 +31,7 @@ while True:
             respone = 'deleted'
         elif action == "l":
             response = list_secrets()
-            
+
         elif action == "x":
             break
         else:
